backend/requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
langchain==0.1.10
langchain_community==0.0.25
langchain_core==0.1.28
openai==1.13.3
pandas==1.5.3
playsound==1.3.0
python-dotenv==1.0.1
ragas==0.1.3
torch==2.2.0
tqdm==4.66.1
TTS==0.22.0
sentence-transformers==2.5.1
pypdf==4.0.2
faiss-cpu==1.8.0
